E2E shell script fixed and optimized

+++ scripts/e2e.test.sh
@@ -2,6 +2,9 @@
 
 clear
 
+
+echo "Checking port $SERVER_PORT" 
+
 # Make sure the port is not already bound 
 if ss -lnt | grep -q :$SERVER_PORT; then 
     echo "Another process is already listening to port $SERVER_PORT" 
@@ -9,9 +12,14 @@ if ss -lnt | grep -q :$SERVER_PORT; then
 fi
 
 # Make sure we are creating clean images
-docker rm -f $(docker ps -a -q)
+if [ ! -z "$(docker ps -a -q)" ]
+then
+    echo "Cleaning docker images." 
+    docker rm -f $(docker ps -a -q)
+fi
 
 # Set the environment (Database, MQTT, etc)
+echo "Cleaning docker-compose stack." 
 docker-compose -f ./scripts/docker-compose.yml up -d
 
 # Wait until mongo logs that it's ready (or timeout after 60s)
@@ -23,9 +31,10 @@ while !(nc -z localhost 27017) && [[ $COUNTER -lt 60 ]] ; do
 done
 
 # TEST Mongo: nc -v localhost 27017
-echo "Connected in  ($COUNTER seconds so far)"
+echo "Connected in ($COUNTER seconds so far)"
 
 # Start the server (production-mode) in background
+echo "Starting the server (production-mode) in background" 
 npm run start &
 
 # Test if the server is up and running
@@ -35,10 +44,15 @@ until ss -lnt | grep -q :$SERVER_PORT; do
 done
 
 # Run the End-to-End tests
+echo "Starting end to end tests" 
 npx dotenv -e .env cucumber-js -- ./test/e2e/features/**/*.feature --require-module ts-node/register --require ./test/e2e/steps/**/*.steps.ts
 
-# Kill everyting SIGTERM signal
-kill -15 0
-
 # Tear down
-docker-compose down
+echo "Destroying docker-compose stack." 
+#docker-compose down
+docker-compose -f ./scripts/docker-compose.yml down
+
+
+# Kill everyting SIGTERM signal
+echo "Killing group processes" 
+kill -15 0
